third_party/WebKit/LayoutTests/editing/deleting/delete_reply_test.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
    [
        '<div contenteditable>',
            '<div>br</div>',
            '<div>',
                '<blockquote type="cite">',
                    '<div>foo</div>',
                    '<div>bar</div>',
                    '<div><br></div>',
                '</blockquote>',
                '<div>|<br></div>',
            '</div>',
        '</div>',
    ].join(''),
    selection => {
        selection.document.execCommand('delete');
        selection.document.execCommand('delete');
    },
    [
        '<div contenteditable>',
            '<div>br</div>',
            '<div>',
                '<blockquote type="cite">',
                    '<div>foo</div>',
                    '<div>bar|</div>',
                '</blockquote>',
            '</div>',
        '</div>',
    ].join('')),
    'Delete to BR in BLOCKQUOTE');

test(() => assert_selection(
    [
        '<div contenteditable>',
            '<div>br</div>',
            '<div>',
                '<blockquote type="cite">',
                    '<div>foo</div>',
                    '<div>bar|</div>',
                '</blockquote>',
            '</div>',
        '</div>',
    ].join(''),
    'insertNewlineInQuotedContent',
    [
        '<div contenteditable>',
            '<div>br</div>',
            '<div>',
                '<blockquote type="cite">',
                    '<div>foo</div>',
                    '<div>bar</div>',
                '</blockquote>',
                '|<br>',
            '</div>',
        '</div>',
    ].join('')),
    'Insert newline in BLOCKQUOTE');
</script>
